chore(refactor): Better directory naming
