+++ src/main/java/controllers/viewController/MonopolyViewController.java
@@ -1,4 +1,4 @@
-package controllers.view;
+package controllers.viewController;
 
 import controllers.ButtonClickNotifier;
 import exception.UnknownGameException;
